add: `Proto3.Suite.DotProto.AST.Core` module
